# Use stackalloc in string.Split