--- a/src/mscorlib/shared/System/Collections/Generic/ValueListBuilder.cs
+++ b/src/mscorlib/shared/System/Collections/Generic/ValueListBuilder.cs
@@ -1,7 +1,12 @@
-﻿using System.Diagnostics;
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System.Buffers;
+using System.Diagnostics;
 using System.Runtime.CompilerServices;
 
-namespace System.Buffers
+namespace System.Collections.Generic
 {
     internal ref struct ValueListBuilder<T>
     {
@@ -40,6 +45,7 @@ namespace System.Buffers
             if (_arrayFromPool != null)
             {
                 ArrayPool<T>.Shared.Return(_arrayFromPool);
+                _arrayFromPool = null;
             }
         }