@AndyAyersMS 2017-12-08
Do we have data on the distribution of string lengths seen by split?

If very short strings are common we might special case the 32 bytes or less case too, as current jit will optimize these further (#14623). If you can show the distribution peaks at a value not too much larger than 32 we could perhaps increase the threshold to extend the range of the jit optimization. I left it at 32 because larger values started causing local offset code bloat and I did not know of any cases that would benefit.