@cod7alex 2017-12-12
@jkotas I made changes only for char parameter version now, am i moving in right direction? 
Benchmarks:
```
|             Method |       Mean |     Error |    StdDev |     Gen 0 |  Allocated |
|------------------- |-----------:|----------:|----------:|----------:|-----------:|
|  SplitCharLength20 |   198.8 us |  1.642 us |  1.536 us |   68.6035 |  210.94 KB |
| SplitCharLength200 | 1,131.7 us | 10.376 us |  9.706 us |  501.9531 | 1546.88 KB |
| SplitCharLength600 | 3,217.5 us | 22.944 us | 21.461 us | 1468.7500 | 4515.63 KB |
```

Looked a bit at string versions and they seem to be not so easy to change due to logic around `lengthList`.