@jkotas 2017-12-12
@cod7alex Yes, I think you are on the right path.