@danmosemsft 2017-12-13
>Do we have data on the distribution of string lengths seen by split?

@billwert for this question