@cod7alex 2017-12-16
Latest version has two problematic benchmarks, see below. Not sure what is the reason to be honest.

| Method                                   | Mean after | Mean before | Mean diff (%) | Allocated after | Allocated before | Allocated diff (%) | 
|------------------------------------------|------------|-------------|---------------|-----------------|------------------|--------------------| 
| SplitCharLength20                        | 2.023 ms   | 2.082 ms    | 2.83          | 2109.38 KB      | 3.05 MB          | 32.46              | 
| SplitCharLength200                       | 11.159 ms  | 11.388 ms   | 2.01          | 15468.75 KB     | 22.96 MB         | 34.21              | 
| SplitCharLength600                       | 31.034 ms  | 31.345 ms   | 0.99          | 45156.25 KB     | 67.21 MB         | 34.39              | 
| SplitStringLength20                      | 1.934 ms   | 1.814 ms    | -6.62         | 1328.13 KB      | 2.29 MB          | 43.36              | 
| SplitStringLength200                     | 7.912 ms   | 8.231 ms    | 3.88          | 8359.38 KB      | 16.02 MB         | 49.04              | 
| SplitStringLength600                     | 20.456 ms  | 23.353 ms   | 12.41         | 23984.38 KB     | 46.54 MB         | 49.67              | 
| SplitStringLength600RemoveEmpty          | 21.132 ms  | 23.097 ms   | 8.51          | 26640.63 KB     | 49.13 MB         | 47.05              | 
| SplitStringArrayLength20                 | 3.330 ms   | 3.466 ms    | 3.92          | 2109.38 KB      | 4.04 MB          | 49.01              | 
| SplitStringArrayLength200                | 17.739 ms  | 18.109 ms   | 2.04          | 11953.13 KB     | 27.39 MB         | 57.38              | 
| SplitStringArrayLength600                | 49.494 ms  | 52.430 ms   | 5.60          | 33828.13 KB     | 79.27 MB         | 58.33              | 
| SplitStringArrayLength600RemoveEmpty     | 49.959 ms  | 50.931 ms   | 1.91          | 38828.13 KB     | 84.15 MB         | 54.94              | 
| SplitCharLength1000NoMatches             | 6.073 ms   | 7.319 ms    | 17.02         | 312.5 KB        | 38.68 MB         | 99.21              | 
| SplitCharLength1000AllCharsAreSeparators | 69.571 ms  | 63.715 ms   | -9.19         | 78437.5 KB      | 114.97 MB        | 33.37              |