@jkotas 2017-12-16
> Not sure what is the reason to be honest

This is allocation heavy benchmark. It is not unusual to see fluctuations like these in allocation heavy benchmarks. The GC kicks in at random spots. These random spots change as you change the allocation pattern. It may result into slowing down innocent victims.

To prove this theory, you can try running the individual tests in a different order, or changing between workstation and server GC.