@cod7alex 2017-12-18
Case with 1000 separators is still slower with suggested experiments. PerfView CPU stacks: 
before:

Name | Exc % | Exc | Inc % | Inc
-- | -- | -- | -- | --
stringsplitbenchmarks!StringSplitBenchmarks.Program.Main() | 0,0 | 6 | 98,8 | 21 769
system.private.corelib.il!String.SplitInternal | 0,3 | 74 | 98,8 | 21 763
system.private.corelib.il!String.SplitKeepEmptyEntries | 0,4 | 86 | 98,3 | 21 654
system.private.corelib.il!System.String.Substring(int32,int32) | 0,4 | 81 | 97,8 | 21 545

after:

Name | Exc % | Exc | Inc % | Inc
-- | -- | -- | -- | --
stringsplitbenchmarks!StringSplitBenchmarks.Program.Main() | 0,0 | 3 | 98,9 | 23 317
system.private.corelib.il!String.SplitInternal | 0,0 | 2 | 98,8 | 23 306
system.private.corelib.il!String.SplitKeepEmptyEntries | 0,1 | 26 | 93,2 | 21 972
system.private.corelib.il!System.String.Substring(int32,int32) | 0,2 | 52 | 92,5 | 21 823
system.private.corelib.il!String.MakeSeparatorList | 0,1 | 20 | 5,3 | 1 246
system.private.corelib.il!System.Buffers.ValueListBuilder`1[System.Int32].Grow() | 0,0 | 1 | 5,1 | 1 201
system.private.corelib.il!System.Span`1[System.Int32].TryCopyTo(value class System.Span`1) | 0,0 | 0 | 1,9 | 448
system.private.corelib.il!System.Buffers.ConfigurableArrayPool`1[System.Int32].Rent(int32) | 0,7 | 170 | 1,4 | 330
system.private.corelib.il!System.Buffers.ConfigurableArrayPool`1[System.Int32].Return(!0[],bool) | 0,8 | 190 | 1,4 | 330
system.private.corelib.il!System.Span.CopyTo(!!0&,!!0&,int32) | 0,8 | 177 | 1,3 | 307