@jkotas 2017-12-19
The problem is that the MakeSeparatorList is 2x slower for SplitCharLength1000AllCharsAreSeparators with this change. I do not think that the copy is the problem. The problem seems to be in the ValueListBuilder.Append method. This method is causing the hot loop in MakeSeparatorList to have about 2x more instructions.

Try to tweak the implementation of this method. E.g. making it look like the similar method in ValueStringBuilder may help: https://github.com/dotnet/coreclr/blob/6cee3ea2010cf754c8f4667bfbc75e06b1f06159/src/mscorlib/shared/System/Text/ValueStringBuilder.cs#L63

Also, changing `if (pos < _span.Length)` to `if ((uint)pos < (uint)_span.Length)` may help because it should allow JIT to eliminate the bounds check.