@cod7alex 2017-12-25
@jkotas Sorry, my results were for some weird test version of code, missed that. The reason is truly Append, and changing it helps to get better results, but it is still slower than before (btw uint trick does not change anything).

In my opinion, it is not possible to achieve same performance for this test case (`SplitCharLength1000AllCharsAreSeparators`) with this approach since we can not eliminate the `if` statement, which is additional logic on each append comparing to array insert.

What i think is possible to do is to slightly tweak first version of PR so it gets stackalloc`ed Span or array from pool depending on string length instead of just allocation of array for big strings.