@jkotas 2017-12-25
> it is still slower than before

How much slower? I think it is acceptable to take some regression for this corner case with this change.