@cod7alex 2017-12-26
Actually, I don`t see the improvement while running clean benchmark with just that test, sorry about confusion once more. Moving to the same layout as list builder has same results (~8% slower) or there are very small diffs like 1ms.