@jkotas 2017-12-26
I think it is fine to take the 8% regression for the case of long strings that are all separators here.

Could you please address the other feedback? It looks fine to me otherwise.